tests: Move service-account helpers to a separate file
